kaggle/AnimalCLEF25/requirements.txt
torch
torchvision
pandas
timm
pyyaml
Pillow
